shared/termios: Remove requirement on cgo when setting raw mode
