--- shared/termios/cgo.go
@@ -1,12 +1,0 @@
-//go:build !windows && cgo
-// +build !windows,cgo
-
-package termios
-
-// #cgo CFLAGS: -std=gnu11 -Wvla -Werror -fvisibility=hidden -Winit-self
-// #cgo CFLAGS: -Wformat=2 -Wshadow -Wendif-labels -fasynchronous-unwind-tables
-// #cgo CFLAGS: -pipe --param=ssp-buffer-size=4 -g -Wunused
-// #cgo CFLAGS: -Werror=implicit-function-declaration
-// #cgo CFLAGS: -Werror=return-type -Wendif-labels -Werror=overflow
-// #cgo CFLAGS: -Wnested-externs -fexceptions
-import "C"
