--- shared/termios/termios_bsd.go
@@ -1,10 +1,0 @@
-//go:build darwin || dragonfly || freebsd || netbsd || openbsd
-// +build darwin dragonfly freebsd netbsd openbsd
-
-package termios
-
-import (
-	"golang.org/x/sys/unix"
-)
-
-const ioctlReadTermios = unix.TIOCGETA
